spatial::hash_grid_Xd_t::hash_grid_Xd_t(): n_containers is now a u32 instead of a usize

+++ gsx/src/internal_spatial/hash_grid_2d.h
@@ -20,7 +20,7 @@ namespace gsx::spatial
     class hash_grid_2d_t : public base_container_2d_t<T>
     {
     public:
-        hash_grid_2d_t(math::vec2 cell_size, usize n_containers)
+        hash_grid_2d_t(math::vec2 cell_size, u32 n_containers)
             : _cell_size(cell_size)
         {
             if (math::min_component(cell_size) <= 0.0f)
